+++ src/editor/init.ts
@@ -2,9 +2,9 @@
 // Import the module and reference it with the alias vscode in your code below
 import * as vscode from 'vscode'
 
+import { deactivate as deactivateMachine } from '../state'
 import createCommands from './commands'
 import createViews from './views'
-import createStateMachine from '../state'
 
 // this method is called when your extension is activated
 // your extension is activated the very first time the command is executed
@@ -14,13 +14,11 @@ export function activate(context: vscode.ExtensionContext) {
     // commands
     createCommands(context)
 
-    // tasks
-    // add tasks here
-
     // views
     createViews(context)
 
-
+    // tasks
+    // add tasks here
 }
 
 // this method is called when your extension is deactivated
@@ -30,4 +28,6 @@ export function deactivate(context: vscode.ExtensionContext): void {
     for (const disposable of context.subscriptions) {
         disposable.dispose()
     }
+    // shut down state machine
+    deactivateMachine()
 }
