+++ src/state/index.ts
@@ -1,28 +1,31 @@
 import { interpret } from 'xstate'
 import machine from './machine'
 
+const machineOptions = {
+    logger: console.log,
+    devTools: true,
+    deferEvents: true,
+    execute: true
+}
+// machine interpreter
+// https://xstate.js.org/docs/guides/interpretation.html
+const service = interpret(machine, machineOptions)
+    // logging
+    .onTransition(state => {
+        console.log('state', state)
+        if (state.changed) {
+            console.log('transition')
+            console.log(state.value)
+        }
+    })
 
-const createStateMachine = () => {
-    const machineOptions = {
-        logger: console.log,
-        devTools: true,
-        deferEvents: true,
-        execute: true
-    }
-    // machine interpreter
-    // https://xstate.js.org/docs/guides/interpretation.html
-    const service = interpret(machine, machineOptions)
-        // logging
-        .onTransition(state => {
-            console.log('state', state)
-            if (state.changed) {
-                console.log('transition')
-                console.log(state.value)
-            }
-        })
+export function activate() {
     // initialize
     service.start()
-    return service
 }
 
-export default createStateMachine
+export function deactivate() {
+    service.stop()
+}
+
+export default service
